third_party/blink/web_tests/external/wpt/svg/animations/svglength-animation-invalid-value-2-expected.txt

This is a testharness.js-based test.
[FAIL] Test SVGLength animation with invalid value: No spaces before value.
  assert_approx_equals: expected 100 +/- 1 but got 0
Harness: the test ran to completion.
